pyproject.toml
[project]
name = "STARK"
version = "0.1.0"
description = "AI-driven data engineering pipeline and multi-agent orchestration system designed to optimize athletic performance"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "duckdb>=1.4.4",
    "openai>=2.21.0",
    "pandas>=3.0.1",
    "plotly>=6.5.2",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "pydantic-ai>=0.0.36",
    "pydantic-ai-slim[web]>=1.67.0",
    "garminconnect>=0.3.1",
    "fitparse>=1.2.0",
    "pyarrow>=19.0.0",
    "playwright>=1.58.0",
    "streamlit>=1.56.0",
]

[tool.uv.workspace]
members = [
    "report-demo",
]
